Added Payment Functionality

+++ backend/tickethub/Repositories/Interfaces/IUserRepository.cs
@@ -3,12 +3,12 @@
     public interface IUserRepository
     {
         Task<List<User>> GetAsync();
-        Task<User> GetByIdAsync(int id);
-        Task<User> AuthenticateAsync(String email, String password);
+        Task<User> GetByIdAsync(long id);
+        Task<User> AuthenticateAsync(String email, String password);   
         Task<User> ValidateEmailAsync(String email);
         Task<User> ValidatePhoneAsync(String phone);
         Task AddAsync(User user);
         Task UpdateAsync(User user);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
