+++ backend/tickethub/Services/Interfaces/IUserService.cs
@@ -5,12 +5,12 @@ namespace tickethub.Services.Interfaces
     public interface IUserService
     {
         Task<List<User>> GetAsync();
-        Task<User> GetByIdAsync(int id);
+        Task<User> GetByIdAsync(long id);
         Task<AuthenticateResponse> AuthenticateAsync(String email, String password);
         Task<User> ValidateEmailAsync(String email);
         Task<User> ValidatePhoneAsync(String phone);
         Task AddAsync(User user);
-        Task UpdateAsync(User user);
-        Task DeleteAsync(int id);
+        Task UpdateAsync(User user);  
+        Task DeleteAsync(long id);
     }
 }
